% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{discrim_mixture}
\alias{discrim_mixture}
\title{Mixture discriminant analysis}
\usage{
discrim_mixture(mode = "classification", sub_classes = NULL)
}
\description{
\code{discrim_mixture} defines an LDA extension model that estimates within-class
density using mixture distributions.
}
