% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{tidymda-package}
\alias{tidymda}
\alias{tidymda-package}
\title{tidymda: Mixture discriminant analysis in tidymodels}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Qiushi Yan \email{qiushi.yann@gmail.com}

}
\keyword{internal}
