% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_mixture.R
\name{discrim_mixture}
\alias{discrim_mixture}
\title{Mixture discriminant analysis}
\usage{
discrim_mixture(mode = "classification", sub_classes = NULL)
}
\arguments{
\item{mode}{A single character string fort the model type. Only possible value
'classification'.}

\item{sub_classes}{An integer for the number of sub-classes in each class.}
}
\description{
\code{discrim_mixture} defines an LDA extension model that estimates within-class
density using mixture distributions.
}
